+++ packages/core/src/index.ts
@@ -41,3 +41,4 @@ export * from "./hooks/UseOnClickOutside";
 export * from "./hooks/UseOnNoMouseInput";
 
 export * from "./utils/SwitchCaseExhauster";
+export * from "./utils/parsers/NumberParser";